[TASK] Rename previously created PSR-14 events
New namings:

* `ApacheSolrForTypo3\Solr\Event\Indexing\AfterIndexItemEvent` is now called `ApacheSolrForTypo3\Solr\Event\Indexing\AfterItemHasBeenIndexedEvent`
* `ApacheSolrForTypo3\Solr\Event\Indexing\AfterIndexItemsEvent` is now called `ApacheSolrForTypo3\Solr\Event\Indexing\AfterItemsHaveBeenIndexedEvent`
* `ApacheSolrForTypo3\Solr\Event\Indexing\BeforeIndexItemEvent` is now called `ApacheSolrForTypo3\Solr\Event\Indexing\BeforeItemIsIndexedEvent`
* `ApacheSolrForTypo3\Solr\Event\Indexing\BeforeIndexItemsEvent` is now called `ApacheSolrForTypo3\Solr\Event\Indexing\BeforeItemsAreIndexedEvent`
* `ApacheSolrForTypo3\Solr\Event\Parser\AfterFacetParsedEvent` is now called `ApacheSolrForTypo3\Solr\Event\Parser\AfterFacetIsParsedEvent`
* `ApacheSolrForTypo3\Solr\Event\Routing\PostProcessUriEvent` is now called `ApacheSolrForTypo3\Solr\Event\Routing\AfterUriIsProcessedEvent`
* `ApacheSolrForTypo3\Solr\Event\Routing\BeforeProcessCachedVariablesEvent` is now called `ApacheSolrForTypo3\Solr\Event\Routing\BeforeCachedVariablesAreProcessedEvent`
* `ApacheSolrForTypo3\Solr\Event\Routing\BeforeReplaceVariableInCachedUrlEvent` is now called `ApacheSolrForTypo3\Solr\Event\Routing\BeforeVariableInCachedUrlAreReplacedEvent`
* `ApacheSolrForTypo3\Solr\Event\Search\AfterFrequentlySearchedEvent` is now called `ApacheSolrForTypo3\Solr\Event\Search\AfterFrequentlySearchHasBeenExecutedEvent`
* `ApacheSolrForTypo3\Solr\Event\Search\AfterSearchEvent` is now called `ApacheSolrForTypo3\Solr\Event\Search\BeforeSearchResultIsShownEvent`
* `ApacheSolrForTypo3\Solr\Event\Search\FormEvent` is now called `ApacheSolrForTypo3\Solr\Event\Search\BeforeSearchFormIsShownEvent`

Documentation has been updated

Fixes: #3681
